fix(react): clear doctor review findings
